feat: sidebar component of chat along with test
